+++ GlobalAccountsService/Repositories/IAccountsRepository.cs
@@ -6,7 +6,7 @@ namespace GlobalAccountsService.Repositories
     {
         Account GetAccountById(String accountId);
         IEnumerable<Account> GetAccounts();
-        void TransferFunds(String fromAccount, String toAccount, decimal amount);
+        void TransferFunds(String fromAccount, String toAccount, float amount);
         void ReactivateAccount(String accountId);
         void CloseAccount(String accountId);
     }
